refactor: 테스트 코드 패키징 변경
