+++ src/test/java/com/mobble/mobbleserver/infrastructure/persistence/member/MemberRepositoryTest.java
@@ -1,9 +1,8 @@
-package com.mobble.mobbleserver.refactor.member.repository;
+package com.mobble.mobbleserver.infrastructure.persistence.member;
 
 import com.mobble.mobbleserver.account.auth.oauth.service.SocialProvider;
 import com.mobble.mobbleserver.config.QueryDslConfig;
 import com.mobble.mobbleserver.domain.member.Member;
-import com.mobble.mobbleserver.infrastructure.persistence.member.JpaMemberRepository;
 import com.mobble.mobbleserver.support.fixture.member.MemberTestFixture;
 import jakarta.persistence.EntityManager;
 import org.junit.jupiter.api.DisplayName;
